refactor: use waitFor
