+++ src/test/suite/oidc.test.ts
@@ -9,12 +9,12 @@ import * as vscode from 'vscode';
 import { createHash } from 'crypto';
 import { before, after, afterEach, beforeEach } from 'mocha';
 import EventEmitter, { once } from 'events';
-import util from 'util';
 import { ExtensionContextStub } from './stubs';
 import { StorageController } from '../../storage';
 import { TelemetryService } from '../../telemetry';
 import ConnectionController from '../../connectionController';
 import { StatusView } from '../../views';
+import { waitFor } from './waitFor';
 
 import { MongoCluster } from 'mongodb-runner';
 import type { MongoClusterOptions } from 'mongodb-runner';
@@ -24,10 +24,6 @@ import { ConnectionString } from 'mongodb-connection-string-url';
 
 import launchMongoShell from '../../commands/launchMongoShell';
 import { getFullRange } from './suggestTestHelpers';
-
-const sleep = (ms: number): Promise<void> => {
-  return util.promisify(setTimeout)(ms);
-};
 
 chai.use(chaiAsPromised);
 
@@ -405,7 +401,9 @@ suite('OIDC Tests', function () {
     }
 
     await reAuthPromise;
-    await sleep(100);
+    await waitFor(() => {
+      return testConnectionController.isCurrentlyConnected() === false;
+    }, 100);
 
     // Because we declined the auth in showInformationMessage above
     expect(tokenFetchCalls).to.equal(1);
